Add a `scan_and_merge` API
Most users of this library are going to be using e.g. serde
to parse the individual config files into a final merged state.

This API is similar to e.g.
https://doc.rust-lang.org/std/iter/trait.Iterator.html#method.reduce

As the API comment says - right now this is not more efficient,
but it could be if we unified things internally because we could
avoid allocating a `BTreeMap` of all the filenames.