Add a `const SYSTEMD_CONVENTIONAL_BASES`
I noticed that the toplevel example in the code has `/etc`
override `/run`, but the systemd-style convention is the opposite.
Let's add a handy `const` which gets this right.  I'll change the
docs in a future PR to use this, and hopefully other projects will
match this and not make a subtle mistake.